requirements.txt
attrs==24.2.0
base58==2.1.1
bitarray==2.9.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cytoolz==0.12.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
hexbytes==0.3.1
httplib2==0.22.0
idna==3.8
ipfshttpclient==0.7.0a1
jsonschema==3.2.0
lru-dict==1.3.0
multiaddr==0.0.9
netaddr==1.3.0
oauthlib==3.2.2
parsimonious==0.8.1
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pyparsing==3.1.4
pyrsistent==0.20.0
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
requests-oauthlib==2.0.0
rlp==2.0.1
rsa==4.9
six==1.16.0
toolz==0.12.1
uritemplate==4.1.1
urllib3==2.2.2
varint==1.0.2
web3==5.19.0
websockets==8.1
